### 2024-01-12 #6225 @editorialbot
[PRE REVIEW]: GrainLearning: A Bayesian uncertainty quantification toolbox for discrete and continuum numerical models of granular materials
**Submitting author:** @luisaforozco (<a href="http://orcid.org/0000-0002-9153-650X">Luisa Fernanda Orozco</a>)
**Repository:** https://github.com/GrainLearning/grainLearning
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v2.0.2
