### 2024-01-15 6fc30c4
Remove class attributes and set_curr_step function -related to #55
