### 2024-01-15 7f338b4
Remove class attributes
