+++ grainlearning/dynamic_systems.py
@@ -87,9 +87,12 @@ class DynamicSystem:
 
     :param obs_data: observation or reference data
     :param num_samples: Sample size of the ensemble of model evaluations
+    :param num_steps: Number of steps or sequence size in the dataset
+    :param num_obs:  Number of observations in the dataset
+    :param num_ctrl: Number of control data (identical between simulation and observation)
     :param param_min: List of parameter lower bounds
     :param param_max: List of parameter Upper bounds
-    :param callback: Callback function, defaults to None
+    :param callback: Callback function to run simulations, defaults to None
     :param curr_iter: current iteration ID, defaults to 0
     :param ctrl_data: control data (e.g, time), defaults to None, optional
     :param obs_names: Column names of the observation data, defaults to None, optional
@@ -101,91 +104,11 @@ class DynamicSystem:
     :param sigma_max: Maximum uncertainty, defaults to 1.0e6, optional
     :param sigma_tol: Tolerance of the estimated uncertainty, defaults to 1.0e-3, optional
     :param sim_name: Name of the simulation, defaults to 'sim', optional
+    :param sigma_min: Minimum uncertainty, defaults to 1.0e-6, optional
+    :param _inv_normalized_sigma:  Calculated normalized sigma to weigh the covariance matrix
+    :param estimated_params: Estimated parameter as the first moment of the distribution (:math:`x_\mu = \sum_i w_i * x_i`), defaults to None, optional
+    :param estimated_params_cv: Estimated parameter coefficient of variation as the second moment of the distribution (:math:`x_\sigma = \sqrt{\sum_i w_i * (x_i - x_\mu)^2} / x_\mu`), defaults to None, optional
     """
-
-    ##### Parameters #####
-
-    #: Parameter data of shape (num_samples, num_params)
-    param_data: np.ndarray
-
-    #: Parameter data of previous iteration
-    param_data_prev: np.ndarray
-
-    #: Number of unknown parameters
-    num_params: int
-
-    #: Lower bound of the parameters
-    param_min: List
-
-    #: Upper bound of the parameters
-    param_max: List
-
-    #: Names of the parameters
-    param_names: List[str]
-
-    ##### Observations #####
-
-    #: Observation (or reference) data of shape (num_obs, num_steps)
-    obs_data: np.ndarray
-
-    #: Observation keys
-    obs_names: List[str]
-
-    #: Inverse of the observation weight
-    inv_obs_weight: List[float]
-
-    #: Number of steps or sequence size in the dataset
-    num_steps: int
-
-    #: Number of observations in the dataset
-    num_obs: int
-
-    #: Control dataset (num_ctrl, num_steps)
-    ctrl_data: np.ndarray
-
-    #: Observation control (e.g., time)
-    ctrl_name: str
-
-    #: Number of control data (identical between simulation and observation)
-    num_ctrl: int
-
-    ##### Simulations #####
-
-    #: Name of the simulation (e.g., sim)
-    sim_name: str = 'sim'
-
-    #: Simulation data of shape (num_samples, num_obs, num_steps)
-    sim_data: np.ndarray
-
-    #: Number of samples (usually specified by user)
-    num_samples: int
-
-    #: Callback function to run the forward predictions
-    callback: Callable
-
-    #: Current iteration ID
-    curr_iter: int
-
-    ##### Uncertainty #####
-
-    #: Minimum value of the uncertainty
-    sigma_min: float = 1.0e-6
-
-    #: Maximum value of the uncertainty
-    sigma_max: float = 1.0e6
-
-    #: Sigma tolerance
-    sigma_tol: float = 1.0e-3
-
-    #: Calculated normalized sigma to weigh the covariance matrix
[…]
