### 2024-01-15 7df6424
Remove class attributes
