### 2024-01-16 b5f872e
minor changes to default values
