### 2024-01-16 9d342ae
Retiefasuarus/issue55 (#62)
* Remove class attributes and set_curr_step function -related to #55

* Remove class attributes

* Fix issue when plotting obs sims when num_ctrl is small

* Remove class attributes

* minor changes to default values

---------

Co-authored-by: chyalexcheng <chyalexcheng@gmail.com>
