### 2024-01-16 6333bab
fix selftests
