+++ tutorials/physics_based/triaxial_compression/triax_calibration.py
@@ -12,18 +12,19 @@ executable = 'yade-batch'
 yade_script = f'{PATH}/triax_YADE_DEM_model.py'
 
 
-def run_sim(system, **kwargs):
+def run_sim(calib):
     """
     Run the external executable and passes the parameter sample to generate the output file.
     """
     print("*** Running external software YADE ... ***\n")
-    os.system(' '.join([executable, system.param_data_file, yade_script]))
+    os.system(' '.join([executable, calib.system.param_data_file, yade_script]))
 
 
 calibration = BayesianCalibration.from_dict(
     {
         "curr_iter": 0,
         "num_iter": 4,
+        "callback": run_sim,
         "system": {
             "system_type": IODynamicSystem,
             "param_min": [7, 0.0, 0.0, 0.0, 10.0],
@@ -37,7 +38,6 @@ calibration = BayesianCalibration.from_dict(
             "sim_data_dir": PATH + '/sim_data/',
             "sim_data_file_ext": '.txt',
             "sigma_tol": 0.01,
-            "callback": run_sim,
         },
         "calibration": {
             "inference": {"ess_target": 0.3},
