+++ tests/unit/test_dynamic_systems.py
@@ -1,13 +1,7 @@
 """Test the dynamic system module."""
 from os import path
-from shutil import rmtree
-from math import floor, log
 import numpy as np
 from grainlearning import DynamicSystem, IODynamicSystem
-from grainlearning.tools import write_dict_to_file
-
-PATH = path.abspath(path.dirname(__file__))
-GL_PATH = path.abspath(path.join(__file__, "../../.."))
 
 
 class TestDynamicSystem:
@@ -31,7 +25,6 @@ class TestDynamicSystem:
             "obs_data": [[12, 3, 4, 4], [12, 4, 5, 4]],
             "ctrl_data": [1, 2, 3, 4],
             "num_samples": 10,
-            "callback": None,
         }
         system_dct = DynamicSystem.from_dict(config)
 
@@ -46,36 +39,6 @@ class TestDynamicSystem:
                 [0.0, 0.70710678],
             ],
             0.00001,
-        )
-
-    def test_run(self):
-        """Test if the callback function works as expected"""
-
-        def run_sim(system):
-            system.sim_data = [
-                [[12, 3, 4, 4]],
-                [[11, 24, 4, 3]],
-            ]  # must be of shape (num_samples,num_obs,num_steps), thus adding a dummy dimension
-
-        config = {
-            "param_min": [1, 2],
-            "param_max": [3, 4],
-            "obs_data": [2, 4, 6, 7],
-            "ctrl_data": [1, 2, 3, 4],
-            "inv_obs_weigh": [0.5, 0.25],
-            "num_samples": 2,
-            "callback": run_sim,
-        }
-        system_cls = DynamicSystem.from_dict(config)
-
-        system_cls.run_callback()
-
-        np.testing.assert_almost_equal(
-            system_cls.sim_data,
-            [
-                [[12, 3, 4, 4]],
-                [[11, 24, 4, 3]],
-            ],
         )
 
     def test_compute_inv_normalized_sigma(self):
@@ -217,70 +180,6 @@ class TestIODynamicSystem:
 
         np.testing.assert_equal(system_cls.__dict__, system_dct.__dict__)
 
-    def test_run(self):
-        """Test if the callback function works as expected"""
-
-        def run_sim(system):
-            """Run the simulation"""
-            # Initialize the data list
-            data = []
-            # Get the ctrl data
-            x = system.ctrl_data
-            # Get the simulation name
-            sim_name = system.sim_name
-            # Loop over the parameter samples
-            for i, param in enumerate(system.param_data):
-                # Get the description of the current sample
-                mag = floor(log(system.num_samples, 10)) + 1
-                description = 'Iter' + str(system.curr_iter) + '_Sample' + str(i).zfill(mag)
-                # Run the model
-                y = param[0] + param[1] * x + param[2] * x ** 2 + param[3] * x ** 3
-                # Append the data to the list
-                data.append(np.array(y, ndmin=2))
-                # Write the data to a file
-                data_file_name = f'{sim_name}_' + description + '_sim.txt'
-                write_dict_to_file({'f': list(y)}, data_file_name)
-                # Write the parameters to a file
-                data_param_name = f'{sim_name}_' + description + '_param.txt'
-                param_data = {'a': [param[0]], 'b': [param[1]], 'c': [param[2]], 'd': [param[3]]}
-                write_dict_to_file(param_data, data_param_name)
-            # Set the simulation data
-            system.set_sim_data(data)
-
-        system_cls = IODynamicSystem(sim_name='test', sim_data_dir=PATH + '/sim_data/', sim_data_file_ext='.txt',
-                                     obs_data_file=path.abspath(
[…]

[…]
