### 2024-01-21 a332f24
remove callback function from dynamic system class
