+++ paper.bib
@@ -62,7 +62,7 @@ volume = {410},
 pages = {116040},
 year = {2023},
 issn = {0045-7825},
-doi = {https://doi.org/10.1016/j.cma.2023.116040},
+doi = {10.1016/j.cma.2023.116040},
 url = {https://www.sciencedirect.com/science/article/pii/S0045782523001640},
 author = {Guilherme Barros and Victor Sapucaia and Philipp Hartmann and Andre Pereira and Jerzy Rojek and Klaus Thoeni},
 keywords = {Discrete Element Method (DEM), Boundary Element Method (BEM), Discontinuous materials, Wave propagation, Infinite domain, Monolithic coupling},
