+++ README.md
@@ -70,7 +70,7 @@ However, you still need to clone the GrainLearning repository to run the tutoria
 
 1. Linear regression with
    the [`run_sim`](tutorials/simple_regression/linear_regression/python_linear_regression_solve.py#L14)
-   callback function of the [`DynamicSystem`](grainlearning/dynamic_systems.py)
+   callback function of the [`BayesianCalibration`](grainlearning/bayesian_calibration.py)
    class,
    in [python_linear_regression_solve.py](tutorials/simple_regression/linear_regression/python_linear_regression_solve.py)
 
