### 2024-02-01 8da6669
Create a new preprocessor to handle data directly coming from GL. Change naming from load_sequences, contact_parameters, outputs to inputs, params, outputs
