### 2024-02-03 e6644f4
add a nonlinear regression tutorial
