### 2024-02-03 8235eca
add a rnn example with nonlinear function
