### 2024-02-07 f20c5ef
change some naming to make the parameters more general. Update the hyperbola calibration example to allow use of ML surrogate
