+++ grainlearning/rnn/models.py
@@ -17,11 +17,11 @@ def rnn_model(
     """
     Neural network with an LSTM layer.
 
-    Takes in a load sequence and contact parameters, and outputs the macroscopic responses.
-    The contact parameters are used to initialize the hidden state of the LSTM.
+    Takes in an input sequence and the parameters and produces an output sequence.
+    The parameters are used to initialize the hidden state of the LSTM.
 
     :param input_shapes: Dictionary containing `'num_input_features'`, `'num_params'`,
-        `'num_labels'`. It can contain other keys but hese are the ones used here.
+        `'num_labels'`. It can contain other keys but these are the ones used here.
     :param window_size: Length of time window.
     :param lstm_units: Number of units of the hidden state of the LSTM.
     :param dense_units: Number of units used in the dense layer after the LSTM.
@@ -52,3 +52,27 @@ def rnn_model(
     model = Model(inputs=[load_sequence, contact_params], outputs=outputs)
 
     return model
+
+
+def rnn_model_for_triax(
+        input_shapes: dict,
+        window_size: int = 20,
+        lstm_units: int = 50,
+        dense_units: int = 20,
+        seed: int = 42,
+        **_,
+        ):
+    """
+    A wrapper of neural network model with an LSTM layer for triaxial loading conditions.
+    :param input_shapes: Dictionary containing `'num_load_features'`, `'num_contact_params'`, and `'num_labels'`.
+    :param window_size: Length of time window.
+    :param lstm_units: Number of units of the hidden state of the LSTM.
+    :param dense_units: Number of units used in the dense layer after the LSTM.
+    :param seed: The random seed used to initialize the weights.
+
+    :return: A Keras model.
+    """
+    # change the name of the keys to match the original model
+    input_shapes['num_input_features'] = input_shapes.pop('num_load_features')
+    input_shapes['num_params'] = input_shapes.pop('num_contact_params')
+    return rnn_model(input_shapes, window_size, lstm_units, dense_units, seed)
