+++ grainlearning/rnn/evaluate_model.py
@@ -29,8 +29,7 @@ def plot_predictions(model: tf.keras.Model, data: tf.data.Dataset, train_stats: 
     plt.rcParams['axes.labelsize'] = 25
     plt.rcParams['font.family'] = 'sans-serif'
 
-    predictions = predict.predict_macroscopics(model, data, train_stats, config,
-                                       batch_size=batch_size)
+    predictions = predict.predict_batch(model, data, train_stats, config, batch_size=batch_size)
     # extract tensors from dataset
     test_inputs, labels = next(iter(data.batch(batch_size)))
 
