### 2024-02-07 3da2496
change some naming to make the parameters more general. Update the hyperbola calibration example to allow use of ML surrogate
