### 2024-02-07 54ee30b
implement a function to create and update input data for rnn.predict_batch
