### 2024-02-07 183e42d
implement a function to create and update input data for rnn.predict_batch
