+++ tutorials/data_driven/LSTM/hyperbola_calibration.py
@@ -4,7 +4,6 @@ from grainlearning.rnn import preprocessor
 import numpy as np
 from grainlearning import BayesianCalibration
 from matplotlib import pyplot as plt
-import tensorflow as tf
 
 x_obs = np.arange(100)
 # hyperbola in a form similar to the Duncan-Chang material model, q = \eps / (a * 100 + b * \eps)
@@ -70,8 +69,6 @@ my_config = {
     'output_data': calibration.system.sim_data,
     'train_frac': 0.7,
     'val_frac': 0.2,
-    # TODO: hide `pad_length` from the user
-    'pad_length': 10,
     'window_size': 10,
     'window_step': 1,
     'patience': 25,
@@ -100,8 +97,7 @@ plt.show()
 # 4. Make predictions with the trained model
 model, train_stats, config = predict_rnn.get_pretrained_model('outputs')
 
-data_inputs = ({'inputs': preprocessor.input_data, 'params': calibration.system.param_data}, preprocessor.input_data)
-data_inputs = tf.data.Dataset.from_tensor_slices(data_inputs)
+data_inputs = preprocessor.prepare_input_data(calibration.system.param_data)
 
 predictions = predict_rnn.predict_batch(model, data_inputs, train_stats, config,
                                         batch_size=calibration.system.num_samples)
@@ -120,11 +116,7 @@ plt.show()
 
 # define the callback function using the ML surrogate
 def run_sim_surrogate(calib):
-    # extend the first dimension of the input data to the number of samples (TODO: hide this from the user)
-    input_data = preprocessor.input_data[0, :, :]
-    preprocessor.input_data = np.repeat(input_data[np.newaxis, :, :], calib.system.num_samples, axis=0)
-    data_inputs = ({'inputs': preprocessor.input_data, 'params': calib.system.param_data}, preprocessor.input_data)
-    data_inputs = tf.data.Dataset.from_tensor_slices(data_inputs)
+    data_inputs = preprocessor.prepare_input_data(calib.system.param_data)
     # make predictions with the trained model
     sim_data = predict_rnn.predict_batch(model, data_inputs, train_stats, config, batch_size=calib.system.num_samples)
     # converting the predictions to GL format (temporal dimension at the end)
@@ -133,6 +125,7 @@ def run_sim_surrogate(calib):
     calib.system.set_sim_data(sim_data)
 
 
+# set the callback function to the one that runs the ML surrogate
 calibration.callback = run_sim_surrogate
 
 # continue the calibration with the surrogate
