### 2024-02-07 cc8c57e
first working example of mixed use of data-driven and equation-based models. 1. Generate sim_data for one iteration; 2. train ML; 3. Split param samples into two subsets: one for ML surrogate, one for equation-based; 4. Continue for all remaining iterations
