### 2024-02-07 bdb03ec
rewrite a cleaner version of hyperbola_calibration_mixed.py
