+++ tutorials/data_driven/LSTM/hyperbola_calibration_mixed.py
@@ -10,29 +10,113 @@ x_obs = np.arange(100)
 y_obs = x_obs / (0.2 * 100 + 5.0 * x_obs)
 
 
-def run_sim(calib):
-    """This is the callback function that runs different realizations of the same model.
-
-    :param calib: The calibration object.
-    """
-    data = []
-    for params in calib.system.param_data:
-        # Run the model
-        y_sim = nonlinear(calib.system.ctrl_data, params)
-        data.append(np.array(y_sim, ndmin=2))
-    calib.system.set_sim_data(data)
-
-
 def nonlinear(x, params):
     return x / (params[0] * 100 + params[1] * x)
 
 
+# Define the configuration dictionary for the ML surrogate
+my_config = {
+    'input_data': None,
+    'param_data': None,
+    'output_data': None,
+    'train_frac': 0.7,
+    'val_frac': 0.2,
+    'window_size': 10,
+    'window_step': 1,
+    'patience': 25,
+    'epochs': 20,
+    'learning_rate': 1e-4,
+    'lstm_units': 128,
+    'dense_units': 128,
+    'batch_size': 64,
+    'standardize_outputs': True,
+    'save_weights_only': True
+}
+
+
+def run_sim_original(x, params):
+    """Run different realizations of the original model.
+
+    :param x: the input sequence
+    :param params: the parameters
+    """
+    data = []
+    for params in params:
+        # Run the model
+        y = nonlinear(x, params)
+        data.append(np.array(y, ndmin=2))
+    return np.array(data)
+
+
+def run_sim_surrogate(params_origin, output_origin, params_surrogate):
+    """Train the ML surrogate and evaluate model output with the ML surrogate.
+
+    :param params_origin: The parameter data used by the original model.
+    :param output_origin: The output data produced by the original model.
+    :param params_surrogate: The parameter data to be used by the ML surrogate.
+    """
+    # expend the parameter and output data
+    my_config['param_data'] = np.vstack([my_config['param_data'], params_origin])
+    my_config['output_data'] = np.vstack([my_config['output_data'], output_origin])
+
+    preprocessor_lstm = preprocessor.PreprocessorLSTM.from_dict(my_config)
+    _ = train_rnn.train_without_wandb(preprocessor_lstm, config=my_config)
+    model, train_stats, config = predict_rnn.get_pretrained_model('outputs')
+
+    # run the surrogate for the second half of the samples
+    data_inputs = preprocessor_lstm.prepare_input_data(params_surrogate)
+    # make predictions with the trained model
+    output_surrogate = predict_rnn.predict_batch(model, data_inputs, train_stats, config,
+                                                 batch_size=params_surrogate.shape[0])
+    # converting the predictions to GL format (temporal dimension at the end)
+    output_surrogate = np.moveaxis(output_surrogate, 1, -1)
+
+    return output_surrogate
+
+
+# 3. Define the callback function using the ML surrogate
+def run_sim_mixed(calib):
+    """This is the callback function that runs different realizations of the same model.
+
+    :param calib: The calibration object.
+    """
+    # if first iteration, run the original function
+    if calib.curr_iter == 0:
+        sim_data = run_sim_original(calib.system.ctrl_data, calib.system.param_data)
+        calib.system.set_sim_data(sim_data)
+        my_config['input_data'] = calib.system.ctrl_data
+        my_config['param_data'] = calib.system.param_data
+        my_config['output_data'] = calib.system.sim_data
+    else:
+        # split samples into two subsets to be used with the original function and the ML surrogate
+        np.random.seed()
[…]
