### 2024-02-08 76b24fc
add a parameter to the sampling module to allow user defining the threshold for slice sampling
