### 2024-02-09 99cff85
Correct broken links in rnn/README
