### 2024-02-15 5101085
add a correct example of hyperparameter tuning using wandb
