### 2024-02-15 ca9053d
move hypertuning into the rnn.train module
