### 2024-02-15 da27040
move hypertuning into the rnn.train module
