### 2024-02-15 2016333
update the rnn tutorial
