### 2024-02-15 e0f1474
update the rnn tutorial
