### 2024-02-15 276c256
setting hypersearch range must be done after setting the default values of other config parameters.
