### 2024-02-19 5b92bd6
fix examples of parameter sweep using wandb
