+++ grainlearning/rnn/train.py
@@ -15,7 +15,7 @@ from grainlearning.rnn.preprocessor import Preprocessor
 from grainlearning.rnn.windows import windowize_single_dataset
 
 
-def train(preprocessor: Preprocessor, config = None, model: tf.keras.Model = None):
+def train(preprocessor: Preprocessor, config=None, model: tf.keras.Model = None):
     """
     Train a model and report to weights and biases.
 
@@ -38,6 +38,7 @@ def train(preprocessor: Preprocessor, config = None, model: tf.keras.Model = Non
         config = wandb.config
         config = _check_config(config, preprocessor)
         config_optimizer = _get_optimizer_config(config)
+        preprocessor.run_dir = wandb.run.dir
 
         # preprocess data
         split_data, train_stats = preprocessor.prepare_datasets()
@@ -49,10 +50,10 @@ def train(preprocessor: Preprocessor, config = None, model: tf.keras.Model = Non
 
         optimizer = tf.keras.optimizers.Adam(**config_optimizer)
         model.compile(
-                optimizer=optimizer,
-                loss='mse',
-                metrics=['mae'],
-            )
+            optimizer=optimizer,
+            loss='mse',
+            metrics=['mae'],
+        )
 
         # create batches
         for split in ['train', 'val']:  # do not batch test set
@@ -60,25 +61,25 @@ def train(preprocessor: Preprocessor, config = None, model: tf.keras.Model = Non
 
         # set up training
         early_stopping = tf.keras.callbacks.EarlyStopping(
-                monitor='val_loss',
-                patience=config.patience,
-                restore_best_weights=True,
-            )
+            monitor='val_loss',
+            patience=config.patience,
+            restore_best_weights=True,
+        )
         wandb_callback = wandb.keras.WandbCallback(
-                monitor='val_loss',
-                save_model=True,
-                save_weights_only=config.save_weights_only,
-                validation_data=split_data['val'],
-            )
+            monitor='val_loss',
+            save_model=True,
+            save_weights_only=config.save_weights_only,
+            validation_data=split_data['val'],
+        )
         callbacks = [wandb_callback, early_stopping]
 
         # train
         history = model.fit(
-                split_data['train'],
-                epochs=config.epochs,
-                validation_data=split_data['val'],
-                callbacks=callbacks,
-            )
+            split_data['train'],
+            epochs=config.epochs,
+            validation_data=split_data['val'],
+            callbacks=callbacks,
+        )
 
         # Evaluate in test dataset and log to wandb the metrics
         test_data = windowize_single_dataset(split_data['test'], **config)
@@ -88,7 +89,8 @@ def train(preprocessor: Preprocessor, config = None, model: tf.keras.Model = Non
 
         return history
 
-def train_without_wandb(preprocessor: Preprocessor, config = None, model: tf.keras.Model = None):
+
+def train_without_wandb(preprocessor: Preprocessor, config=None, model: tf.keras.Model = None):
     """
     Train a model locally: no report to wandb.
     Saves either the model or its weight to folder outputs.
@@ -108,7 +110,8 @@ def train_without_wandb(preprocessor: Preprocessor, config = None, model: tf.ker
     if os.path.exists(path_save_data):
         delete_outputs = input(f"The contents of {path_save_data} will be permanently deleted,\
                                  do you want to proceed? [y/n]: ")
-        if delete_outputs == "y": shutil.rmtree(path_save_data)
+        if delete_outputs == "y":
+            shutil.rmtree(path_save_data)
         else:
             raise SystemExit("Cancelling training")
 
@@ -116,8 +119,8 @@ def train_without_wandb(preprocessor: Preprocessor, config = None, model: tf.ker
 
[…]
