### 2024-02-19 1a12235
fix examples of parameter sweep using wandb
