### 2024-02-19 e85175d
adapt rnn documentation
