+++ tutorials/data_driven/LSTM/hyperbola_calibration_mixed_hypertuning.py
@@ -1,0 +1,196 @@
+import grainlearning.rnn.train as train_rnn
+import grainlearning.rnn.predict as predict_rnn
+from grainlearning.rnn import preprocessor
+import numpy as np
+from grainlearning import BayesianCalibration
+import wandb
+
+x_obs = np.arange(100)
+# hyperbola in a form similar to the Duncan-Chang material model, q = \eps / (a * 100 + b * \eps)
+y_obs = x_obs / (0.2 * 100 + 5.0 * x_obs)
+
+
+def nonlinear(x, params):
+    return x / (params[0] * 100 + params[1] * x)
+
+
+# Define the configuration dictionary for the ML surrogate
+my_config = {
+    'input_data': None,
+    'param_data': None,
+    'output_data': None,
+    'train_frac': 0.7,
+    'val_frac': 0.2,
+    'window_size': 10,
+    'window_step': 1,
+    'patience': 25,
+    'epochs': 20,
+    'learning_rate': 1e-4,
+    'lstm_units': 128,
+    'dense_units': 128,
+    'batch_size': 64,
+    'standardize_outputs': True,
+    'save_weights_only': True
+}
+
+# 2. Define the sweep configuration
+sweep_config = {
+    'method': 'random',
+    'metric': {'goal': 'minimize', 'name': 'mae'},
+    'early_terminate': {
+        'type': 'hyperband',
+        's': 2,
+        'eta': 3,
+        'max_iter': 27
+    }
+}
+
+search_space = {
+    'learning_rate': {
+        # a flat distribution between 1e-4 and 1e-2
+        'distribution': 'q_log_uniform_values',
+        'q': 1e-4,
+        'min': 1e-4,
+        'max': 1e-2
+    },
+    'lstm_units': {
+        'distribution': 'q_log_uniform_values',
+        'q': 1,
+        'min': 32,
+        'max': 256
+    },
+    'window_size': {
+        'distribution': 'q_uniform',
+        'q': 2,
+        'min': 4,
+        'max': 30
+    },
+}
+
+
+def run_sim_original(x, params):
+    """Run different realizations of the original model.
+
+    :param x: the input sequence
+    :param params: the parameters
+    """
+    data = []
+    for params in params:
+        # Run the model
+        y = nonlinear(x, params)
+        data.append(np.array(y, ndmin=2))
+    return np.array(data)
+
+
+def my_training_function():
+    preprocessor_lstm = preprocessor.PreprocessorLSTM.from_dict(my_config)
+    train_rnn.train(preprocessor_lstm, config=my_config)
+
+
+def hyper_train():
+    """Train the ML surrogate using hyperparameter tuning."""
+    hyper_tuner = train_rnn.HyperTuning(sweep_config, search_space, my_config, project_name='hyperbola_sweep')
+    hyper_tuner.run_sweep(my_training_function, count=10)
+
+    hyper_tuner.sweep_config['metric']['name'] = 'val_loss'
+    entity_project_sweep_id = f"{hyper_tuner.entity_name}/{hyper_tuner.project_name}/{hyper_tuner.sweep_id}"
+    model, train_stats, config = predict_rnn.get_best_run_from_sweep(entity_project_sweep_id)
+    return model, train_stats, config
+
+
+def run_sim_surrogate(params_origin, output_origin, params_surrogate):
+    """Train the ML surrogate and evaluate model output with the ML surrogate.
+
+    :param params_origin: The parameter data used by the original model.
+    :param output_origin: The output data produced by the original model.
+    :param params_surrogate: The parameter data to be used by the ML surrogate.
+    """
+    # expend the parameter and output data
+    my_config['param_data'] = np.vstack([my_config['param_data'], params_origin])
[…]
