### 2024-02-20 662f715
resolve issue #71 by excluding input, parameter, and output arrays from the sweep_config of wandb.sweep. However, somehow the arrays get turned into string when reloaded from best_run.config
