### 2024-03-01 ce9af25
add normalization factor to sim data and obs data
