### 2024-03-05 GrainLearning/grainLearning#72 @gchure
JOSS Review: Paper Comments
Hi! I’m [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/6338) and opening small issues as I come across them. This issue corresponds to my comments on the manuscript.

The paper hits all of the desired points by JOSS, but I personally found it to be very terse and not informative about the software. I think substantially more detail should be added to the paper such that, when cited by researchers, a reader can briefly look at the paper and understand the functionality and design of the software. I feel comfortable requesting this change as you have very extensive documentation and I feel that using some of that content to formally present the software in the paper is not too arduous of a task.

Below are my  comments that need to be addressed in the revised paper

1. While the summary does a good job at setting the context of the field, I find the summary to be lacking in describing what GrainLearning does. Could you please add another two sentences or so highlighting the broad strokes of what GrainLearning does, beyond “emulating granular behavior” and “learning the uncertainties?”
2. The ‘State of the Field’ section reads more like what I would expect for the ‘Statement of Need’. I think you should merge these and shorten.
3.  I don’t believe lines 32 - 40 should be in the statement of need as they describe the functionality of the software.
4. The “functionality” section of the paper is highly tailored for a narrow scientific community, and I think obscures the very cool capabilities of the model. I would greatly appreciate if, at a minimum, you could provide a software architecture diagram that highlights the GarinLearning data model. This will allow a quick  reader to glance at the paper and understand what GrainLearning takes in and what it puts out.
5. It is not clear to me what  the authors mean by “calibration” in their functionality statement. This may be because I am not in-field, but a definition would be useful. My reading of this statement does not jibe with the type of model calibration I am used to in Bayesian model evaluation.
6. I also would appreciate a statement of contribution  of the authors. I’m very happy that all authors are listed as co-first authors, but I think with this many, a statement of who-did-what would be useful. If all authors equally contributed to writing/coding/testing/deploying/etc, great! Having that written down would only make  it more clear that this was a highly collaborative effort.
