### 2024-03-12 GrainLearning/grainLearning#72 @luisaforozco
@gchure Could you give us an example of how in JOSS paper we could explicitly mention the different contributions of the different authors? This is not something that we are used to do in the articles of our scientific domain ☺️
