### 2024-03-15 GrainLearning/grainLearning#72 @gchure
Of course. I recently requested this for [Taweret](https://github.com/openjournals/joss-reviews/issues/6175#issuecomment-1916173144), another package in review at JOSS. [I have a paper](https://www.cell.com/fulltext/S2405-4712(18)30057-7) (though not in JOSS) with 5 co-first authors and I felt that the "Author Contribution" section accurately conveyed why we split authorship equally. It can be as simple as "X, Y, and Z developed the software. A, B, X, and Y developed statistical models. X and Z wrote test suites. X wrote the documentation." etc. My interest for this is more to make it really clear to the reader that everyone substantially contributed to the software, which I have found to be really well engineered.
