### 2024-03-15 #6338 @gchure
Hi @diehlpk, I've finished my review. I've found the software to be in very good shape, with no substantive comments on the structure or functionality. Installation works as advertised, though it seems that building the [wheel for Windows](https://github.com/GrainLearning/grainLearning/actions/runs/7862270782/job/21451551279) is currently failing. I don't have access to a windows machine to test installation, however, so maybe this is not an issue (@luisaforozco can maybe comment on this).

My only blocking issue is the status of the software paper (see [GrainLearning-#72](https://github.com/GrainLearning/grainLearning/issues/72)). I think it could use some more work to better represent the capability of the software. As the authors have very nice documentation, I don't think this should take too much effort. Once they've addressed this point, I can mark my review as complete.

Thanks for the patience -- the flu followed by APS march meeting isn't a combo I would recommend, especially if you have reviews to get to!
