### 2024-03-15 #6338 @luisaforozco
Thanks a lot for your review @diehlpk. Regarding the building problem for windows the issue is with a dependency of tensorflow: `tensorflow-io-gcs-filesystem` for which the wheels for windows have been somehow forgotten or not published anymore ([see issue(s)](https://github.com/tensorflow/io/issues/1815) and [reports via other channels](https://discuss.tensorflow.org/t/tensorflow-io-gcs-filesystem-with-windows/18849/4)).  For a local installation, if the users handled the tensorflow installation on their side prior to installing GrainLearning, everything should work as intended.
