+++ paper.md
@@ -35,9 +35,10 @@ bibliography: paper.bib
 How to keep dikes safe with rising sea levels? Why are ripples formed in sand? What can we prepare for landing on Mars? At the center of these questions is the understanding of how the grains, as a self-organizing material, collide, flow, or get jammed and compressed. State-of-the-art algorithms allow for simulating millions of grains individually in a computer. However, such computations can take very long and produce complex data difficult to interpret and be upscaled to large-scale applications such as sediment transport and debris flows. GrainLearning is an open-source toolbox with machine learning and statistical inference modules allowing for emulating granular material behavior and learning material uncertainties from real-life observations.
 
 To understand what GrainLearning does, let us consider a mechanical test performed on a granular material. The macroscopic response of such material, in terms of stress-strain evolution curves, is obtained from the test.
-It would be interesting to have a digital equivalent material to further investigate, using numerical simulations such as the discrete element method (DEM), how such material would behave under other mechanical constraints. To do so, the first step is defining a contact model governing interactions between grains in DEM. Then, using GrainLearning one can calibrate or infer the values of the contact model parameters such that the mechanical response observed in the real-world experiment is the closest to the one obtained in the DEM simulation.
+It would be interesting to have a digital equivalent material to further investigate, using numerical simulations such as the discrete element method (DEM), how such material would behave under other mechanical constraints. To do so, the first step is defining a contact model governing interactions between grains in DEM. This involves multiple a-priori unknown constants, such as friction coefficients or Young's modulus, whose chosen values will determine the macroscopic behavior of the simulation.
+By repeatedly comparing the simulation results with provided experimental data, GrainLearning allows one to calibrate or infer these values such that the mechanical response in the DEM simulation is the closest to that observed in the real-world experiment.
 
-Making abstraction of this idea, GrainLearning's use can be extended to other kind of models or *Dynamical systems* that can be other simulation frameworks such as FEM, CFD, LBM, and even other techniques such as agent based modelling. In the same vein, the framework is not exclusive for granular materials.
+While it was initially developed for DEM simulations of granular materials, GrainLearning can be extended to other simulation frameworks such as FEM, CFD, LBM, and even other techniques such as agent-based modeling. In the same vein, the framework is not exclusive for granular materials.
 
 # Statement of need
 
