### 2024-03-18 GrainLearning/grainLearning#72 @luisaforozco
Dear @gchure, we have worked on the text and added the section *Author contributions*.
You can find the latest version generated in the action [here](https://github.com/GrainLearning/grainLearning/actions/runs/8327789966).
Let us know what you think and thanks again for your comments.
