### 2024-03-18 e40e8d8
Update python versions supported, linter pyproject.toml
