+++ pyproject.toml
@@ -17,7 +17,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.12"
+python = ">=3.8,<=3.12"
 numpy = "^1.23.2"
 scipy = "^1.9.1"
 scikit-learn = "^1.1.3"
@@ -27,14 +27,14 @@ seaborn = '^0.12.2'
 Sphinx = {version = "^5.1.1", optional = true}
 sphinx-autodoc-typehints = {version = "^1.19.2", optional = true}
 sphinx-mdinclude = {version = "*", optional = true}
-sphinx-rtd-theme = {version=">=1.1.1", optional = true}
+sphinx-rtd-theme = {version = ">=1.1.1", optional = true}
 pytest = {version = "^6.2.4", optional = true}
 pytest-cov = {version = "^2.12.1", optional = true}
 prospector = {version = "^1.7.6", optional = true, extras = ["with_pyroma"]}
 pyroma = {version = "^4.0", optional = true}
-h5py = {version ="^3.7.0", optional = true}
-wandb = {version ="^0.13.4", optional = true}
-tensorflow = {version ="^2.10.0", optional = true}
+h5py = {version = "^3.7.0", optional = true}
+wandb = {version = "^0.13.4", optional = true}
+tensorflow = {version = "^2.10.0", optional = true}
 ipykernel = {version = "*", optional = true}
 
 [tool.poetry.extras]
