### 2024-03-18 10956a3
Add note to docs: install via pip on windows (due to tensorflow dependency)
