+++ pyproject.toml
@@ -17,7 +17,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<=3.12"
+python = ">=3.9,<3.12"
 numpy = "^1.23.2"
 scipy = "^1.9.1"
 scikit-learn = "^1.1.3"
