### 2024-03-18 d929f60
Bump wand version after: https://github.com/wandb/wandb/issues/5844
