+++ pyproject.toml
@@ -33,7 +33,7 @@ pytest-cov = {version = "^2.12.1", optional = true}
 prospector = {version = "^1.7.6", optional = true, extras = ["with_pyroma"]}
 pyroma = {version = "^4.0", optional = true}
 h5py = {version = "^3.7.0", optional = true}
-wandb = {version = "^0.13.4", optional = true}
+wandb = {version = "^0.15.10", optional = true}
 tensorflow = {version = "^2.10.0", optional = true}
 ipykernel = {version = "*", optional = true}
 
