### 2024-03-18 GrainLearning/grainLearning#72 @gchure
Looks great. Just a few small comments:

i) Can you reference Figure 1 in the text? I think something like "The core functionality of GrainLearning is diagrammed in Figure 1." at line 58.

ii) The figure caption lists steps 1, 2, 3... but those aren't shown in the figure. This makes it hard to see what each step is referring to in the figure. Can you add them as sub labels within the figure?

With those two small points addressed, I can finish off my review.
