### 2024-03-18 #6225 @Haipeng-ustc
In my review of this repo, I found it is a well-developed software. The installation process works smoothly, although I didn't test on Windows. The code quality is good, and the tutorial provided makes it accessible to its users. Furthermore, the unit test suite is very developed as well. Overall, "GrainLearning" represents a decent package for its related research. I recommend it highly for its utility and design.
