### 2024-03-19 c8e414c
Add labels to image and reference in text.
