### 2024-03-19 GrainLearning/grainLearning#72 @gchure
Great! Looks perfect. Thanks for the rapid iteration!

As a minor point, you'll probably want to switch out the * and ¶ in the author affiliations. [I discovered empirically](https://github.com/openjournals/joss-reviews/issues/6270#issuecomment-1936791995) that this will break your paper upon acceptance.
