### 2024-03-20 f2e581a
remove equal_contrib, statement Author Contributions was added
