### 2024-04-19 #6338 @luisaforozco
Hi @diehlpk, sure, I will tick the boxes.
I have one question: is it expected that the branch containing the files `paper.md` and `paper.bib` woudl be merged to main and the release would include those?
Thanks!
