### 2024-04-19 b8ac771
Update poetry config for readthedocs
