+++ .readthedocs.yaml
@@ -8,10 +8,5 @@ build:
     post_create_environment:
       # install poetry https://docs.readthedocs.io/en/stable/build-customization.html#id11
       - pip install poetry
-      # Tell poetry to not use a virtual environment 
-      # (We are in a virtual environement anyway)
-      - poetry config virtualenvs.create false
     post_install:
-      # commented out because poetry in readthedocs environement was not able to find right packages otherwise
-      # - poetry lock --no-update
       - poetry install -E docs -E rnn
