+++ .readthedocs.yaml
@@ -9,4 +9,7 @@ build:
       # install poetry https://docs.readthedocs.io/en/stable/build-customization.html#id11
       - pip install poetry
     post_install:
-      - poetry install -E docs -E rnn
+      # Install dependencies with 'docs' dependency group
+      # VIRTUAL_ENV needs to be set manually for now.
+      # See https://github.com/readthedocs/readthedocs.org/pull/11152/
+      - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install -E docs -E rnn
