### 2024-04-19 75a03ae
Add Klaus and Philipp as authors of the software paper
