+++ .github/workflows/release.yml
@@ -9,6 +9,7 @@ on:
 jobs:
   build:
     steps:
+      - runs-on: ubuntu-latest
       - uses: actions/checkout@v3
       - name: Set up Python 3.10
         uses: actions/setup-python@v3
