+++ .github/workflows/release.yml
@@ -8,8 +8,8 @@ on:
       
 jobs:
   build:
+    runs-on: ubuntu-latest
     steps:
-      - runs-on: ubuntu-latest
       - uses: actions/checkout@v3
       - name: Set up Python 3.10
         uses: actions/setup-python@v3
