### 2024-04-22 #6338 @luisaforozco
GrainLearning version 2.0.3 has been released.
The zenodo DOI of this release is [10.5281/zenodo.11001174](https://doi.org/10.5281/zenodo.11001174)
_Note:_ the paper should be re-generated to include two extra authors.

Here is my checklist (completed):
> @luisaforozco could you please do the following
>
> * [x]     Double check authors and affiliations (including ORCIDs)
> * [x]    Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x]     Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [x]      Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]      Make sure that the license listed for the archive is the same as the software license.
