### 2024-04-23 #6338 @diehlpk
@openjournals/dev

> Hello @diehlpk, here are the things you can ask me to do:
>
> ```
>
> # List all available commands
> @editorialbot commands
>
> # Add to this issue's reviewers list
> @editorialbot add @username as reviewer
>
> # Remove from this issue's reviewers list
> @editorialbot remove @username from reviewers
>
> # Get a list of all editors's GitHub handles
> @editorialbot list editors
>
> # Assign a user as the editor of this submission
> @editorialbot assign @username as editor
>
> # Remove the editor assigned to this submission
> @editorialbot remove editor
>
> # Remind an author, a reviewer or the editor to return to a review after a
> # certain period of time (supported units days and weeks)
> @editorialbot remind @reviewer in 2 weeks
>
> # Adds a checklist for the reviewer using this command
> @editorialbot generate my checklist
>
> # Set a value for version
> @editorialbot set v1.0.0 as version
>
> # Set a value for branch
> @editorialbot set joss-paper as branch
>
> # Set a value for repository
> @editorialbot set https://github.com/organization/repo as repository
>
> # Set a value for the archive DOI
> @editorialbot set set 10.5281/zenodo.6861996 as archive
>
> # Mention the EiCs for the correct track
> @editorialbot ping track-eic
>
> # Run checks and provide information on the repository and the paper file
> @editorialbot check repository
>
> # Check the references of the paper for missing DOIs
> @editorialbot check references
>
> # Generates the pdf paper
> @editorialbot generate pdf
>
> # Recommends the submission for acceptance
> @editorialbot recommend-accept
>
> # Generates a LaTeX preprint file
> @editorialbot generate preprint
>
> # Flag submission with questionable scope
> @editorialbot query scope
>
> # Get a link to the complete list of reviewers
> @editorialbot list reviewers
>
> # Creates a post-review checklist with editor and authors tasks
> @editorialbot create post-review checklist
>
> # Open the review issue
> @editorialbot start review
> ```

@openjournals/dev the `set set` above looks strange. Can you please have a look?
