### 2024-04-23 #6338 @diehlpk
@luisaforozco The title on Zendo 'GrainLearning' does not match with the paper 'GrainLearning: A Bayesian uncertainty quantification toolbox for discrete and continuum numerical models of granular materials'.

Can you please update the Title? There is no need for a new DOI.
