### 2024-05-01 #6338 @Kevin-Mattheus-Moerman
- [ ] Consider adding the town for the first affiliation.
- [ ] Please spell out `AU` as `Australia` in the last your.
- [ ] `gaussian`  and `bayesian` should start with a capital letter.
