### 2024-05-02 28d8b53
Add cities of affiliations, upper case for Bayesian and Gaussian
