### 2024-05-02 #6338 @luisaforozco
@Kevin-Mattheus-Moerman thanks for the comments! I've made the requested changes to the text.

> @luisaforozco as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below are some points that require your attention:
>
> * [x]  Consider adding the town/city for the first and second affiliation.
> * [x]  Please spell out `AU` as `Australia` in the last your.
> * [x]  `gaussian`  and `bayesian` should start with a capital letter.
