### 2024-05-08 #6338 @luisaforozco
Dear @Kevin-Mattheus-Moerman, a co-author noticed that there was an error in the acknowledgements:
“~The last author~ _H. Cheng_ would like to thank the Netherlands eScience Center for the funding provided under grant number NLESC.OEC.2021.032.”

I've updated the `paper.md` accordingly, is it possible to update the published paper?
Thanks!
