### 2024-05-08 #6338 @Kevin-Mattheus-Moerman
@openjournals/dev not sure what to make of this error? Could you advise and perhaps process the re-accept? Thanks!
