### 2024-05-08 #6338 @xuanxu
The draft generation fails because the paper has already been accepted. You can run the reaccept command directly.
